+++ pkg/lbaas/frontend/definition.go
@@ -1,6 +1,6 @@
 package frontend
 
-import "github.com/anexia-it/go-anxcloud/pkg/lbas/common"
+import "github.com/anexia-it/go-anxcloud/pkg/lbaas/common"
 
 type Definition struct {
 	Name           string       `json:"name"`
